refactor: extract ChatSettingsManager from Config.cs verbatim
